Rework the file structure of the repository
